+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
         tools:targetApi="31">
 
         <service
-            android:name=".data.ClodyFirebaseMessagingService"
+            android:name=".ClodyFirebaseMessagingService"
             android:exported="false">
             <intent-filter>
                 <action android:name="com.google.firebase.MESSAGING_EVENT" />
