+++ app/src/main/java/com/sopt/clody/ClodyFirebaseMessagingService.kt
@@ -1,4 +1,4 @@
-package com.sopt.clody.data
+package com.sopt.clody
 
 import android.content.Context
 import com.google.firebase.messaging.FirebaseMessagingService
